refactor(sandbox): 重构沙箱 API，简化方法调用并删除冗余代码
