+++ examples/sandbox_create/main.go
@@ -45,7 +45,7 @@ func main() {
 	// 演示完毕，终止沙箱释放资源
 	killCtx, killCancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer killCancel()
-	if err := c.Kill(killCtx, sb.ID()); err != nil {
+	if err := sb.Kill(killCtx); err != nil {
 		log.Printf("终止沙箱失败: %v", err)
 	} else {
 		fmt.Printf("沙箱 %s 已终止\n", sb.ID())
